server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "replay-backend"
version = "0.1.0"
description = "FastAPI backend for the Replay meeting note-taker MVP"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.27,<0.29",
    "SQLAlchemy>=2.0,<3.0",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "python-multipart>=0.0.9,<0.1",
    "email-validator>=2.1,<3.0",
    "alembic>=1.12,<2.0",
    "aiofiles>=23.2,<24.0",
    "httpx>=0.27,<0.28",
    "tenacity>=8.2,<9.0",
    "anyio>=4.3,<5.0",
    "typing-extensions>=4.9,<5.0",
    "aiosqlite>=0.19,<0.20",
    "asyncpg>=0.29,<1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8.0",
    "pytest-asyncio>=0.23,<0.24",
    "pytest-cov>=4.1,<5.0",
    "factory-boy>=3.3,<4.0",
    "respx>=0.20,<0.21",
    "ruff>=0.3,<0.4",
    "mypy>=1.8,<2.0"
]

[tool.setuptools.packages.find]
where = ["src"]
